configs/stm32mp157_dk: switch to a TF-A based boot chain
The boot chain used until now by the stm32mp157_dk defconfig was using
U-Boot SPL as the first stage bootloader. However, the
STM32CubeProgrammer utility only works with a boot chain based on TF-A
(ARM Trusted Firmware). So this commit switches the stm32mp157_dk
defconfig to use TF-A as the first stage bootloader.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>